requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==3.2.6
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
sqlparse==0.4.4
typing_extensions==4.7.1
django-heroku==0.3.1
python-decouple==3.8
drf-yasg==1.21.7
django-cors-headers==4.2.0